data aug

+++ src/data/nuscene_dataset_centernet.py
@@ -98,11 +98,11 @@ class NusceneDatasetCenterNet(NusceneDataset):
                         velocity_target[y, x, :] = self.gen_velocity(box["velocity"])
 
         return {
-            "category": torch.FloatTensor(category_target),
-            "attribute": torch.FloatTensor(attribute_target),
-            "offset": torch.FloatTensor(offset_target),
-            "depth": torch.FloatTensor(depth_target),
-            "size": torch.FloatTensor(size_target),
-            "rotation": torch.FloatTensor(rotation_target),
-            "velocity": torch.FloatTensor(velocity_target),
+            "category": category_target,
+            "attribute": attribute_target,
+            "offset": offset_target,
+            "depth": depth_target,
+            "size": size_target,
+            "rotation": rotation_target,
+            "velocity": velocity_target,
         }
